+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 2017-08-15
 
+- exploring some stuff ... before adding Var/Vars within the Vm model... 
+  https://github.com/OlivierBlanvillain/monadic-html/issues/57
 - LLMap to interface with Leaflet
 - make ss1.demo.sh generate deterministic timestamps and values
 - some better organization